+++ project/src/main/java/lu/uni/adtool/ADToolMain.java
@@ -218,6 +218,10 @@ public final class ADToolMain extends JFrame {
     return detailsView;
   }
 
+  public MainController getController() {
+    return this.controller;
+  }
+
   /**
    * Status bar
    */