+++ project/src/main/java/lu/uni/adtool/tree/XmlConverter.java
@@ -36,8 +36,15 @@ import java.io.BufferedOutputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStreamWriter;
+import java.io.PrintWriter;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import bibliothek.util.xml.XElement;
 import bibliothek.util.xml.XException;
@@ -56,7 +63,7 @@ public class XmlConverter {
    * @param idToExport - set od domain IDs to export
    *
    * @throws IOException
-   */
+  */
   public void exportTo(FileOutputStream fileStream, TreeLayout layout, Set<Integer> idToExport) //, ArrayList<ValuationDomain> domains)
       throws IOException {
     XElement rootXML = null;
@@ -134,4 +141,98 @@ public class XmlConverter {
       controller.reportError(Options.getMsg("error.xmlimport.fail") +  e.getMessage());
     }
   }
+
+  /**
+   * Exports tree to the file stream in Latex format.
+   *
+   * @param fileStream - file stream to which we want to export the tree in XML format
+   * @param layout - layout containing the tree and the associated domains
+   *
+   * @throws IOException
+   */
+  public void exportLatex(FileOutputStream fileStream, TreeLayout layout) throws IOException {
+    InputStream headerStream = getClass().getResourceAsStream("/latexHeader.tex");
+    if (headerStream != null) {
+      Debug.log("Exporting latex 2");
+      String header = TxtImporter.read(headerStream);
+      if (layout.getSwitchRole()) {
+        header = header.replaceAll("%%STYLE%%", "  opp/.style={draw=red, attacker, ispro=false, common, inner sep=2pt},"
+                                   + System.getProperty("line.separator")
+                                   + "  pro/.style={draw=green, defender, ispro=true, common, inner sep=1pt},");
+      }
+      else {
+        header = header.replaceAll("%%STYLE%%", "  pro/.style={draw=red, attacker, ispro=true, common, inner sep=2pt},"
+                                   + System.getProperty("line.separator")
+                                   + "  opp/.style={draw=green, defender,ispro=false, common, inner sep=1pt},");
+      }
+      StringBuilder packages = new StringBuilder();
+      String tree = "";
+      if (layout.isSand())  {
+//         tree = ((SandNode)layout.getRoot()).toLatex(0);
+        Debug.log("Exporting sand");
+      }
+      else {
+        tree = ((ADTNode)layout.getRoot()).toLatex(0);
+        Debug.log("Exporting adt" + tree);
+      }
+      if (tree.indexOf("\\textquotesingle") != -1) {
+        packages.append("\\usepackage{textcomp}" + System.getProperty("line.separator"));
+      }
+      if (tree.indexOf("\\textquotedbl") != -1) {
+        packages.append("\\usepackage[T1]{fontenc}"+ System.getProperty("line.separator"));
+      }
+//       header = header.replaceAll("\n", System.getProperty("line.separator"));
+      header = header.replaceAll("\r", System.getProperty("line.separator"));
+      header = header.replaceAll("%%IMPORTS%%", Matcher.quoteReplacement(packages.toString()));
+      header = header.replaceAll("%%TREE%%\\n", Matcher.quoteReplacement(tree));
+      PrintWriter pstream = new PrintWriter(new OutputStreamWriter(fileStream, StandardCharsets.UTF_8), true);
+      String[] lines = header.split(System.getProperty("line.separator"));
+      for(String line: lines){
+        pstream.println(line);
+      }
+      if (pstream.checkError()) {
+        throw new IOException(Options.getMsg("error.write"));
+      }
+      pstream.flush();
+      pstream.close();
+    }
+    else {
+      throw new IOException(Options.getMsg("error.nolatexheader"));
+    }
+  }
+  public static String escape2Latex(String label) {
+    if (pattern == null) {
+      tokens = new HashMap<String,String>();
+      tokens.put("\\", "\\\\textbackslash{}");
+      tokens.put("{", "\\\\{");
+      tokens.put("}", "\\\\}");
+      tokens.put("$", "\\\\\\$");
+      tokens.put("&", "\\\\&");
+      tokens.put("#", "\\\\#");
+      tokens.put("^", "\\\\textasciicircum{}");
+      tokens.put("_", "\\\\_");
+      tokens.put("~", "\\\\textasciitilde{}");
+      tokens.put("%", "\\\\%");
+      tokens.put(",", "{,}");
+      tokens.put("\n", "\\\\\\\\\n");
+      tokens.put("<", "\\\\textless{}");
+      tokens.put(">", "\\\\textgreater{}");
+      tokens.put("|", "\\\\textbar{}");
+      tokens.put("\"", "\\\\textquotedbl{}");
+      tokens.put("\'", "\\\\textquotesingle{}");
+      tokens.put("`", "\\\\textasciigrave{}");
+      pattern = Pattern.compile("(\\\\|\\||`|\\}|\\^|\"|#|%|&|\'|\\$|\n|,|<|~|>|\\{|_)");
+    }
+    Matcher matcher = pattern.matcher(label);
+    StringBuffer sb = new StringBuffer();
+    while(matcher.find()) {
+      String text = matcher.group(1);
+      matcher.appendReplacement(sb, tokens.get(text));
+    }
+    matcher.appendTail(sb);
+
+    return sb.toString();
+  }
+  private static Pattern pattern = null;
+  private static Map<String,String> tokens = null;
 }